requirements.txt
pillow
numpy
requests
datasets
pandas
tqdm
